M3-426 Refactor : 알림을 전송하는 부분과 알림 데이터를 저장하는 부분의 관심사 분리
